common tokens separated from stdlib
