+++ lib/src/metta/runner/stdlib_minimal/mod.rs
@@ -1,19 +1,19 @@
 #[macro_use]
 pub mod stdlib_math;
 
+#[macro_use]
+pub mod stdlib_common_tokens;
+
 use crate::*;
 use crate::space::*;
 use crate::metta::*;
 use crate::metta::text::Tokenizer;
-use crate::metta::types::{get_atom_types, get_meta_type};
 use crate::common::assert::vec_eq_no_order;
 use crate::common::shared::Shared;
 use crate::metta::text::SExprParser;
 use crate::metta::runner::{Metta, RunContext, ModuleLoader, ResourceKey};
 use crate::metta::runner::string::Str;
 use crate::common::CachingMapper;
-use crate::common::multitrie::MultiTrie;
-use crate::space::grounding::atom_to_trie_key;
 #[cfg(feature = "pkg_mgmt")]
 use crate::metta::runner::{git_catalog::ModuleGitLocation, mod_name_from_url, pkg_mgmt::UpdateMode};
 
@@ -212,93 +212,6 @@ impl CustomExecute for IncludeOp {
     }
 }
 
-/// mod-space! returns the space of a specified module, loading the module if it's not loaded already
-//NOTE: The "impure" '!' denoted in the op atom name is due to the side effect of loading the module.  If
-// we want a side-effect-free version, it could be implemented by calling `RunContext::get_module_by_name`
-// instead of `RunContext::load_module`, but then the user would need to use `register-module!`, `import!`,
-// or some other mechanism to make sure the module is loaded in advance.
-#[derive(Clone, Debug)]
-pub struct ModSpaceOp {
-    //TODO-HACK: This is a terrible horrible ugly hack that should be fixed ASAP
-    context: std::sync::Arc<std::sync::Mutex<Vec<std::sync::Arc<std::sync::Mutex<&'static mut RunContext<'static, 'static>>>>>>,
-}
-
-grounded_op!(ModSpaceOp, "mod-space!");
-
-impl ModSpaceOp {
-    pub fn new(metta: Metta) -> Self {
-        Self{ context: metta.0.context.clone() }
-    }
-}
-
-impl Grounded for ModSpaceOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, ATOM_TYPE_ATOM, rust_type_atom::<DynSpace>()])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for ModSpaceOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        let arg_error = "mod-space! expects a module name argument";
-        let mod_name_atom = args.get(0).ok_or_else(|| ExecError::from(arg_error))?;
-
-        // TODO: replace Symbol by grounded String?
-        let mod_name = match mod_name_atom {
-            Atom::Symbol(mod_name) => mod_name.name(),
-            _ => {return Err(ExecError::from(arg_error))}
-        };
-        let mod_name = strip_quotes(mod_name);
-
-        // Load the module into the runner, or get the ModId if it's already loaded
-        //TODO: Remove this hack to access the RunContext, when it's part of the arguments to `execute`
-        let ctx_ref = self.context.lock().unwrap().last().unwrap().clone();
-        let mut context = ctx_ref.lock().unwrap();
-        let mod_id = context.load_module(mod_name)?;
-
-        let space = Atom::gnd(context.metta().module_space(mod_id));
-        Ok(vec![space])
-    }
-}
-
-/// This operation prints the modules loaded from the top of the runner
-///
-/// NOTE: This is a temporary stop-gap to help MeTTa users inspect which modules they have loaded and
-/// debug module import issues.  Ultimately it probably makes sense to make this information accessible
-/// as a special kind of Space, so that it would be possible to work with it programmatically.
-#[derive(Clone, Debug)]
-pub struct PrintModsOp {
-    metta: Metta
-}
-
-grounded_op!(PrintModsOp, "print-mods!");
-
-impl PrintModsOp {
-    pub fn new(metta: Metta) -> Self {
-        Self{ metta }
-    }
-}
-
-impl Grounded for PrintModsOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, UNIT_TYPE])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for PrintModsOp {
-    fn execute(&self, _args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        self.metta.display_loaded_modules();
-        unit_result()
-    }
-}
-
 #[derive(Clone, Debug)]
 pub struct BindOp {
     tokenizer: Shared<Tokenizer>,
@@ -335,114 +248,6 @@ impl CustomExecute for BindOp {
 }
 
 #[derive(Clone, Debug)]
-pub struct NewSpaceOp {}
-
-grounded_op!(NewSpaceOp, "new-space");
-
-impl Grounded for NewSpaceOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, rust_type_atom::<DynSpace>()])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for NewSpaceOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        if args.len() == 0 {
-            let space = Atom::gnd(DynSpace::new(GroundingSpace::new()));
-            Ok(vec![space])
-        } else {
-            Err("new-space doesn't expect arguments".into())
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct AddAtomOp {}
-
-grounded_op!(AddAtomOp, "add-atom");
-
-impl Grounded for AddAtomOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, rust_type_atom::<DynSpace>(),
-            ATOM_TYPE_ATOM, UNIT_TYPE])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for AddAtomOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        let arg_error = || ExecError::from("add-atom expects two arguments: space and atom");
-        let space = args.get(0).ok_or_else(arg_error)?;
-        let atom = args.get(1).ok_or_else(arg_error)?;
-        let space = Atom::as_gnd::<DynSpace>(space).ok_or("add-atom expects a space as the first argument")?;
-        space.borrow_mut().add(atom.clone());
-        unit_result()
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct RemoveAtomOp {}
-
-grounded_op!(RemoveAtomOp, "remove-atom");
-
-impl Grounded for RemoveAtomOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, rust_type_atom::<DynSpace>(),
-            ATOM_TYPE_ATOM, UNIT_TYPE])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for RemoveAtomOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        let arg_error = || ExecError::from("remove-atom expects two arguments: space and atom");
-        let space = args.get(0).ok_or_else(arg_error)?;
-        let atom = args.get(1).ok_or_else(arg_error)?;
-        let space = Atom::as_gnd::<DynSpace>(space).ok_or("remove-atom expects a space as the first argument")?;
-        space.borrow_mut().remove(atom);
-        // TODO? Is it necessary to distinguish whether the atom was removed or not?
-        unit_result()
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct GetAtomsOp {}
-
-grounded_op!(GetAtomsOp, "get-atoms");
-
-impl Grounded for GetAtomsOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, rust_type_atom::<DynSpace>(),
-            ATOM_TYPE_ATOM])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for GetAtomsOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        let arg_error = || ExecError::from("get-atoms expects one argument: space");
-        let space = args.get(0).ok_or_else(arg_error)?;
-        let space = Atom::as_gnd::<DynSpace>(space).ok_or("get-atoms expects a space as its argument")?;
-        space.borrow().as_space().atom_iter()
-            .map(|iter| iter.cloned().map(|a| make_variables_unique(a)).collect())
-            .ok_or(ExecError::Runtime("Unsupported Operation. Can't traverse atoms in this space".to_string()))
-    }
-}
-
-#[derive(Clone, Debug)]
 pub struct PragmaOp {
     settings: Shared<HashMap<String, Atom>>,
 }
@@ -474,58 +279,6 @@ impl CustomExecute for PragmaOp {
         unit_result()
     }
 }
-
-#[derive(Clone, Debug)]
-pub struct GetTypeSpaceOp {}
-
-grounded_op!(GetTypeSpaceOp, "get-type-space");
-
-impl Grounded for GetTypeSpaceOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, rust_type_atom::<DynSpace>(), ATOM_TYPE_ATOM, ATOM_TYPE_ATOM])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for GetTypeSpaceOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        let arg_error = || ExecError::from("get-type-space expects two arguments: space and atom");
-        let space = args.get(0).ok_or_else(arg_error)?;
-        let space = Atom::as_gnd::<DynSpace>(space).ok_or("get-type-space expects a space as the first argument")?;
-        let atom = args.get(1).ok_or_else(arg_error)?;
-        log::debug!("GetTypeSpaceOp::execute: space: {}, atom: {}", space, atom);
-
-        Ok(get_atom_types(space, atom))
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct GetMetaTypeOp { }
-
-grounded_op!(GetMetaTypeOp, "get-metatype");
-
-impl Grounded for GetMetaTypeOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, ATOM_TYPE_ATOM, ATOM_TYPE_ATOM])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for GetMetaTypeOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        let arg_error = || ExecError::from("get-metatype expects single atom as an argument");
-        let atom = args.get(0).ok_or_else(arg_error)?;
-
-        Ok(vec![get_meta_type(&atom)])
-    }
-}
-
 
 #[derive(Clone, Debug)]
 pub struct PrintlnOp {}
@@ -638,27 +391,6 @@ impl CustomExecute for TraceOp {
     }
 }
 
-#[derive(Clone, Debug)]
-pub struct NopOp {}
-
-grounded_op!(NopOp, "nop");
-
-impl Grounded for NopOp {
-    fn type_(&self) -> Atom {
-        ATOM_TYPE_UNDEFINED
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for NopOp {
-    fn execute(&self, _args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        unit_result()
-    }
-}
-
 #[derive(Clone, PartialEq, Debug)]
 pub struct StateAtom {
     state: Rc<RefCell<Atom>>
@@ -688,79 +420,6 @@ impl Grounded for StateAtom {
             Atom::Grounded(a) => a.type_(),
         };
         Atom::expr([expr!("StateMonad"), typ])
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct NewStateOp { }
-
-grounded_op!(NewStateOp, "new-state");
-
-impl Grounded for NewStateOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, expr!(tnso), expr!("StateMonad" tnso)])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for NewStateOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        let arg_error = "new-state expects single atom as an argument";
-        let atom = args.get(0).ok_or(arg_error)?;
-        Ok(vec![Atom::gnd(StateAtom::new(atom.clone()))])
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct GetStateOp { }
-
-grounded_op!(GetStateOp, "get-state");
-
-impl Grounded for GetStateOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, expr!("StateMonad" tgso), expr!(tgso)])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for GetStateOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        let arg_error = "get-state expects single state atom as an argument";
-        let state = args.get(0).ok_or(arg_error)?;
-        let atom = Atom::as_gnd::<StateAtom>(state).ok_or(arg_error)?;
-        Ok(vec![atom.state.borrow().clone()])
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct ChangeStateOp { }
-
-grounded_op!(ChangeStateOp, "change-state!");
-
-impl Grounded for ChangeStateOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, expr!("StateMonad" tcso), expr!(tcso), expr!("StateMonad" tcso)])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for ChangeStateOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        let arg_error = "change-state! expects a state atom and its new value as arguments";
-        let atom = args.get(0).ok_or(arg_error)?;
-        let state = Atom::as_gnd::<StateAtom>(atom).ok_or("change-state! expects a state as the first argument")?;
-        let new_value = args.get(1).ok_or(arg_error)?;
-        *state.state.borrow_mut() = new_value.clone();
-        Ok(vec![atom.clone()])
     }
 }
 
@@ -826,33 +485,6 @@ impl CustomExecute for EqualOp {
         let b = args.get(1).ok_or_else(arg_error)?;
 
         Ok(vec![Atom::gnd(Bool(a == b))])
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct MatchOp {}
-
-grounded_op!(MatchOp, "match");
-
-impl Grounded for MatchOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, rust_type_atom::<DynSpace>(), ATOM_TYPE_ATOM, ATOM_TYPE_ATOM, ATOM_TYPE_UNDEFINED])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for MatchOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        let arg_error = || ExecError::from("match expects three arguments: space, pattern and template");
-        let space = args.get(0).ok_or_else(arg_error)?;
-        let pattern = args.get(1).ok_or_else(arg_error)?;
-        let template = args.get(2).ok_or_else(arg_error)?;
-        log::debug!("MatchOp::execute: space: {:?}, pattern: {:?}, template: {:?}", space, pattern, template);
-        let space = Atom::as_gnd::<DynSpace>(space).ok_or("match expects a space as the first argument")?;
-        Ok(space.borrow().subst(&pattern, &template))
     }
 }
 
@@ -984,313 +616,6 @@ pub(crate) mod pkg_mgmt_ops {
     }
 }
 
-#[derive(Clone, Debug)]
-pub struct UniqueAtomOp {}
-
-grounded_op!(UniqueAtomOp, "unique-atom");
-
-impl Grounded for UniqueAtomOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, ATOM_TYPE_EXPRESSION, ATOM_TYPE_EXPRESSION])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for UniqueAtomOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        let arg_error = || ExecError::from("unique expects single expression atom as an argument");
-        let expr = TryInto::<&ExpressionAtom>::try_into(args.get(0).ok_or_else(arg_error)?)?;
-
-        let mut atoms: Vec<Atom> = expr.children().into();
-        let mut set = GroundingSpace::new();
-        atoms.retain(|x| {
-            let not_contained = set.query(x).is_empty();
-            if not_contained { set.add(x.clone()) };
-            not_contained
-        });
-        Ok(vec![Atom::expr(atoms)])
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct UnionAtomOp {}
-
-grounded_op!(UnionAtomOp, "union-atom");
-
-impl Grounded for UnionAtomOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, ATOM_TYPE_EXPRESSION, ATOM_TYPE_EXPRESSION, ATOM_TYPE_EXPRESSION])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for UnionAtomOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        let arg_error = || ExecError::from("union expects and executable LHS and RHS atom");
-        let mut lhs: Vec<Atom> = TryInto::<&ExpressionAtom>::try_into(args.get(0).ok_or_else(arg_error)?)?.children().into();
-        let rhs: Vec<Atom> = TryInto::<&ExpressionAtom>::try_into(args.get(1).ok_or_else(arg_error)?)?.children().into();
-
-        lhs.extend(rhs);
-
-        Ok(vec![Atom::expr(lhs)])
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct IntersectionAtomOp {}
-
-grounded_op!(IntersectionAtomOp, "intersection-atom");
-
-impl Grounded for IntersectionAtomOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, ATOM_TYPE_EXPRESSION, ATOM_TYPE_EXPRESSION, ATOM_TYPE_EXPRESSION])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for IntersectionAtomOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        let arg_error = || ExecError::from("intersection expects and executable LHS and RHS atom");
-        let mut lhs: Vec<Atom> = TryInto::<&ExpressionAtom>::try_into(args.get(0).ok_or_else(arg_error)?)?.children().into();
-        let rhs = TryInto::<&ExpressionAtom>::try_into(args.get(1).ok_or_else(arg_error)?)?.children();
-
-        let mut rhs_index: MultiTrie<SymbolAtom, Vec<usize>> = MultiTrie::new();
-        for (index, rhs_item) in rhs.iter().enumerate() {
-            let k = atom_to_trie_key(&rhs_item);
-            // FIXME this should
-            // a) use a mutable value endpoint which the MultiTrie does not support atm
-            // b) use a linked list, which Rust barely supports atm
-            let r = rhs_index.get(&k).next();
-            match r.cloned() {
-                Some(bucket) => {
-                    rhs_index.remove(&k, &bucket);
-                    let mut nbucket = bucket;
-                    nbucket.push(index);
-                    let nbucket = nbucket;
-                    rhs_index.insert(k, nbucket);
-                }
-                None => { rhs_index.insert(k, vec![index]) }
-            }
-        }
-
-        lhs.retain(|candidate| {
-            let k = atom_to_trie_key(candidate);
-            let r = rhs_index.get(&k).next();
-            match r.cloned() {
-                None => { false }
-                Some(bucket) => {
-                    match bucket.iter().position(|item| &rhs[*item] == candidate) {
-                        None => { false }
-                        Some(i) => {
-                            rhs_index.remove(&k, &bucket);
-                            if bucket.len() > 1 {
-                                let mut nbucket = bucket;
-                                nbucket.remove(i);
-                                rhs_index.insert(k, nbucket);
-                            }
-                            true
-                        }
-                    }
-                }
-            }
-        });
-
-        Ok(vec![Atom::expr(lhs)])
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct MaxAtomOp {}
-
-grounded_op!(MaxAtomOp, "max-atom");
-
-impl Grounded for MaxAtomOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, ATOM_TYPE_EXPRESSION, ATOM_TYPE_NUMBER])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for MaxAtomOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        let arg_error = || ExecError::from("max-atom expects one argument: expression");
-        let children = TryInto::<&ExpressionAtom>::try_into(args.get(0).ok_or_else(arg_error)?)?.children();
-        if children.is_empty() {
-            Err(ExecError::from("Empty expression"))
-        } else {
-            children.into_iter().fold(Ok(f64::NEG_INFINITY), |res, x| {
-                match (res, Number::from_atom(x)) {
-                    (res @ Err(_), _) => res,
-                    (_, None) => Err(ExecError::from("Only numbers are allowed in expression")),
-                    (Ok(max), Some(x)) => Ok(f64::max(max, x.into())),
-                }
-            })
-        }.map(|max| vec![Atom::gnd(Number::Float(max))])
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct MinAtomOp {}
-
-grounded_op!(MinAtomOp, "min-atom");
-
-impl Grounded for MinAtomOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, ATOM_TYPE_EXPRESSION, ATOM_TYPE_NUMBER])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for MinAtomOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        let arg_error = || ExecError::from("min-atom expects one argument: expression");
-        let children = TryInto::<&ExpressionAtom>::try_into(args.get(0).ok_or_else(arg_error)?)?.children();
-        if children.is_empty() {
-            Err(ExecError::from("Empty expression"))
-        } else {
-            children.into_iter().fold(Ok(f64::INFINITY), |res, x| {
-                match (res, Number::from_atom(x)) {
-                    (res @ Err(_), _) => res,
-                    (_, None) => Err(ExecError::from("Only numbers are allowed in expression")),
-                    (Ok(min), Some(x)) => Ok(f64::min(min, x.into())),
-                }
-            })
-        }.map(|min| vec![Atom::gnd(Number::Float(min))])
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct SizeAtomOp {}
-
-grounded_op!(SizeAtomOp, "size-atom");
-
-impl Grounded for SizeAtomOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, ATOM_TYPE_EXPRESSION, ATOM_TYPE_NUMBER])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for SizeAtomOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        let arg_error = || ExecError::from("size-atom expects one argument: expression");
-        let children = TryInto::<&ExpressionAtom>::try_into(args.get(0).ok_or_else(arg_error)?)?.children();
-        let size = children.len();
-        Ok(vec![Atom::gnd(Number::Integer(size as i64))])
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct IndexAtomOp {}
-
-grounded_op!(IndexAtomOp, "index-atom");
-
-impl Grounded for IndexAtomOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, ATOM_TYPE_EXPRESSION, ATOM_TYPE_NUMBER, ATOM_TYPE_ATOM])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for IndexAtomOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        let arg_error = || ExecError::from("index-atom expects two arguments: expression and atom");
-        let children = TryInto::<&ExpressionAtom>::try_into(args.get(0).ok_or_else(arg_error)?)?.children();
-        let index = args.get(1).and_then(Number::from_atom).ok_or_else(arg_error)?;
-        match children.get(Into::<i64>::into(index) as usize) {
-            Some(atom) => Ok(vec![atom.clone()]),
-            None => Err(ExecError::from("Index is out of bounds")),
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct SubtractionAtomOp {}
-
-grounded_op!(SubtractionAtomOp, "subtraction-atom");
-
-impl Grounded for SubtractionAtomOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, ATOM_TYPE_EXPRESSION, ATOM_TYPE_EXPRESSION, ATOM_TYPE_EXPRESSION])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for SubtractionAtomOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        let arg_error = || ExecError::from("subtraction expects and executable LHS and RHS atom");
-        let mut lhs: Vec<Atom> = TryInto::<&ExpressionAtom>::try_into(args.get(0).ok_or_else(arg_error)?)?.children().into();
-        let rhs = TryInto::<&ExpressionAtom>::try_into(args.get(1).ok_or_else(arg_error)?)?.children();
-
-        let mut rhs_index: MultiTrie<SymbolAtom, Vec<usize>> = MultiTrie::new();
-        for (index, rhs_item) in rhs.iter().enumerate() {
-            let k = atom_to_trie_key(&rhs_item);
-            // FIXME this should
-            // a) use a mutable value endpoint which the MultiTrie does not support atm
-            // b) use a linked list, which Rust barely supports atm
-            let r = rhs_index.get(&k).next();
-            match r.cloned() {
-                Some(bucket) => {
-                    rhs_index.remove(&k, &bucket);
-                    let mut nbucket = bucket;
-                    nbucket.push(index);
-                    let nbucket = nbucket;
-                    rhs_index.insert(k, nbucket);
-                }
-                None => { rhs_index.insert(k, vec![index]) }
-            }
-        }
-
-        lhs.retain(|candidate| {
-            let k = atom_to_trie_key(candidate);
-            let r = rhs_index.get(&k).next();
-            match r.cloned() {
-                None => { true }
-                Some(bucket) => {
-                    match bucket.iter().position(|item| &rhs[*item] == candidate) {
-                        None => { true }
-                        Some(i) => {
-                            rhs_index.remove(&k, &bucket);
-                            if bucket.len() > 1 {
-                                let mut nbucket = bucket;
-                                nbucket.remove(i);
-                                rhs_index.insert(k, nbucket);
-                            }
-                            false
-                        }
-                    }
-                }
-            }
-        });
-
-        Ok(vec![Atom::expr(lhs)])
-    }
-}
-
 //TODO: In the current version of rand it is possible for rust to hang if range end's value is too
 // big. In future releases (0.9+) of rand signature of sample_single will be changed and it will be
 // possible to use match construction to cover overflow and other errors. So after library will be
@@ -1394,77 +719,8 @@ fn atom_to_string(atom: &Atom) -> String {
         _ => atom.to_string(),
     }
 }
-#[derive(Clone, Debug)]
-pub struct GetTypeOp {
-    space: DynSpace,
-}
 
-grounded_op!(GetTypeOp, "get-type");
 
-impl GetTypeOp {
-    pub fn new(space: DynSpace) -> Self {
-        Self{ space }
-    }
-}
-
-impl Grounded for GetTypeOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, ATOM_TYPE_ATOM, ATOM_TYPE_ATOM])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for GetTypeOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        let arg_error = || ExecError::from("get-type expects single atom as an argument");
-        let atom = args.get(0).ok_or_else(arg_error)?;
-        let space = match args.get(1) {
-            Some(space) => Atom::as_gnd::<DynSpace>(space)
-                .ok_or("match expects a space as the first argument"),
-            None => Ok(&self.space),
-        }?;
-        let types = get_atom_types(space, atom);
-        if types.is_empty() {
-            Ok(vec![EMPTY_SYMBOL])
-        } else {
-            Ok(types)
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct IfEqualOp { }
-
-grounded_op!(IfEqualOp, "if-equal");
-
-impl Grounded for IfEqualOp {
-    fn type_(&self) -> Atom {
-        Atom::expr([ARROW_SYMBOL, ATOM_TYPE_ATOM, ATOM_TYPE_ATOM, ATOM_TYPE_ATOM, ATOM_TYPE_ATOM, ATOM_TYPE_ATOM])
-    }
-
-    fn as_execute(&self) -> Option<&dyn CustomExecute> {
-        Some(self)
-    }
-}
-
-impl CustomExecute for IfEqualOp {
-    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
-        let arg_error = || ExecError::from("if-equal expects <atom> <pattern> <then> <else> as an argument");
-        let atom = args.get(0).ok_or_else(arg_error)?;
-        let pattern = args.get(1).ok_or_else(arg_error)?;
-        let then = args.get(2).ok_or_else(arg_error)?;
-        let else_ = args.get(3).ok_or_else(arg_error)?;
-
-        if crate::matcher::atoms_are_equivalent(atom, pattern) {
-            Ok(vec![then.clone()])
-        } else {
-            Ok(vec![else_.clone()])
-        }
-    }
-}
 
 // TODO: remove hiding errors completely after making it possible passing
 // them to the user
@@ -1740,55 +996,55 @@ impl CustomExecute for CaseOp {
 // to the runner & module state.  https://github.com/trueagi-io/hyperon-experimental/issues/410
 pub fn register_common_tokens(tref: &mut Tokenizer, _tokenizer: Shared<Tokenizer>, space: &DynSpace, metta: &Metta) {
 
-    let get_type_op = Atom::gnd(GetTypeOp::new(space.clone()));
+    let get_type_op = Atom::gnd(stdlib_common_tokens::GetTypeOp::new(space.clone()));
     tref.register_token(regex(r"get-type"), move |_| { get_type_op.clone() });
-    let get_type_space_op = Atom::gnd(GetTypeSpaceOp{});
+    let get_type_space_op = Atom::gnd(stdlib_common_tokens::GetTypeSpaceOp{});
     tref.register_token(regex(r"get-type-space"), move |_| { get_type_space_op.clone() });
-    let get_meta_type_op = Atom::gnd(GetMetaTypeOp{});
+    let get_meta_type_op = Atom::gnd(stdlib_common_tokens::GetMetaTypeOp{});
     tref.register_token(regex(r"get-metatype"), move |_| { get_meta_type_op.clone() });
-    let is_equivalent = Atom::gnd(IfEqualOp{});
+    let is_equivalent = Atom::gnd(stdlib_common_tokens::IfEqualOp{});
     tref.register_token(regex(r"if-equal"), move |_| { is_equivalent.clone() });
-    let new_space_op = Atom::gnd(NewSpaceOp{});
+    let new_space_op = Atom::gnd(stdlib_common_tokens::NewSpaceOp{});
     tref.register_token(regex(r"new-space"), move |_| { new_space_op.clone() });
-    let add_atom_op = Atom::gnd(AddAtomOp{});
+    let add_atom_op = Atom::gnd(stdlib_common_tokens::AddAtomOp{});
     tref.register_token(regex(r"add-atom"), move |_| { add_atom_op.clone() });
-    let remove_atom_op = Atom::gnd(RemoveAtomOp{});
+    let remove_atom_op = Atom::gnd(stdlib_common_tokens::RemoveAtomOp{});
     tref.register_token(regex(r"remove-atom"), move |_| { remove_atom_op.clone() });
-    let get_atoms_op = Atom::gnd(GetAtomsOp{});
+    let get_atoms_op = Atom::gnd(stdlib_common_tokens::GetAtomsOp{});
     tref.register_token(regex(r"get-atoms"), move |_| { get_atoms_op.clone() });
-    let new_state_op = Atom::gnd(NewStateOp{});
+    let new_state_op = Atom::gnd(stdlib_common_tokens::NewStateOp{});
     tref.register_token(regex(r"new-state"), move |_| { new_state_op.clone() });
-    let change_state_op = Atom::gnd(ChangeStateOp{});
+    let change_state_op = Atom::gnd(stdlib_common_tokens::ChangeStateOp{});
     tref.register_token(regex(r"change-state!"), move |_| { change_state_op.clone() });
-    let get_state_op = Atom::gnd(GetStateOp{});
+    let get_state_op = Atom::gnd(stdlib_common_tokens::GetStateOp{});
     tref.register_token(regex(r"get-state"), move |_| { get_state_op.clone() });
-    let nop_op = Atom::gnd(NopOp{});
+    let nop_op = Atom::gnd(stdlib_common_tokens::NopOp{});
     tref.register_token(regex(r"nop"), move |_| { nop_op.clone() });
-    let match_op = Atom::gnd(MatchOp{});
+    let match_op = Atom::gnd(stdlib_common_tokens::MatchOp{});
     tref.register_token(regex(r"match"), move |_| { match_op.clone() });
-    let min_atom_op = Atom::gnd(MinAtomOp{});
+    let min_atom_op = Atom::gnd(stdlib_common_tokens::MinAtomOp{});
     tref.register_token(regex(r"min-atom"), move |_| { min_atom_op.clone() });
-    let max_atom_op = Atom::gnd(MaxAtomOp{});
+    let max_atom_op = Atom::gnd(stdlib_common_tokens::MaxAtomOp{});
     tref.register_token(regex(r"max-atom"), move |_| { max_atom_op.clone() });
-    let size_atom_op = Atom::gnd(SizeAtomOp{});
+    let size_atom_op = Atom::gnd(stdlib_common_tokens::SizeAtomOp{});
     tref.register_token(regex(r"size-atom"), move |_| { size_atom_op.clone() });
-    let index_atom_op = Atom::gnd(IndexAtomOp{});
+    let index_atom_op = Atom::gnd(stdlib_common_tokens::IndexAtomOp{});
     tref.register_token(regex(r"index-atom"), move |_| { index_atom_op.clone() });
     let random_int_op = Atom::gnd(RandomIntOp{});
     tref.register_token(regex(r"random-int"), move |_| { random_int_op.clone() });
     let random_float_op = Atom::gnd(RandomFloatOp{});
     tref.register_token(regex(r"random-float"), move |_| { random_float_op.clone() });
-    let mod_space_op = Atom::gnd(ModSpaceOp::new(metta.clone()));
+    let mod_space_op = Atom::gnd(stdlib_common_tokens::ModSpaceOp::new(metta.clone()));
     tref.register_token(regex(r"mod-space!"), move |_| { mod_space_op.clone() });
-    let print_mods_op = Atom::gnd(PrintModsOp::new(metta.clone()));
+    let print_mods_op = Atom::gnd(stdlib_common_tokens::PrintModsOp::new(metta.clone()));
     tref.register_token(regex(r"print-mods!"), move |_| { print_mods_op.clone() });
-    let unique_op = Atom::gnd(UniqueAtomOp{});
+    let unique_op = Atom::gnd(stdlib_common_tokens::UniqueAtomOp{});
     tref.register_token(regex(r"unique-atom"), move |_| { unique_op.clone() });
-    let subtraction_op = Atom::gnd(SubtractionAtomOp{});
+    let subtraction_op = Atom::gnd(stdlib_common_tokens::SubtractionAtomOp{});
     tref.register_token(regex(r"subtraction-atom"), move |_| { subtraction_op.clone() });
-    let intersection_op = Atom::gnd(IntersectionAtomOp{});
+    let intersection_op = Atom::gnd(stdlib_common_tokens::IntersectionAtomOp{});
     tref.register_token(regex(r"intersection-atom"), move |_| { intersection_op.clone() });
-    let union_op = Atom::gnd(UnionAtomOp{});
+    let union_op = Atom::gnd(stdlib_common_tokens::UnionAtomOp{});
     tref.register_token(regex(r"union-atom"), move |_| { union_op.clone() });
     let pow_math_op = Atom::gnd(stdlib_math::PowMathOp {});
     tref.register_token(regex(r"pow-math"), move |_| { pow_math_op.clone() });
@@ -1974,36 +1230,6 @@ mod tests {
     }
 
     #[test]
-    fn get_type_op() {
-        let space = DynSpace::new(metta_space("
-            (: B Type)
-            (: C Type)
-            (: A B)
-            (: A C)
-        "));
-
-        let get_type_op = GetTypeOp::new(space.clone());
-        assert_eq_no_order!(get_type_op.execute(&mut vec![sym!("A"), expr!({space.clone()})]).unwrap(),
-            vec![sym!("B"), sym!("C")]);
-    }
-
-    #[test]
-    fn get_type_op_non_valid_atom() {
-        let space = DynSpace::new(metta_space("
-            (: f (-> Number String))
-            (: 42 Number)
-            (: \"test\" String)
-        "));
-
-        let get_type_op = GetTypeOp::new(space.clone());
-        assert_eq_no_order!(get_type_op.execute(&mut vec![expr!("f" "42"), expr!({space.clone()})]).unwrap(),
-            vec![sym!("String")]);
-        assert_eq_no_order!(get_type_op.execute(&mut vec![expr!("f" "\"test\""), expr!({space.clone()})]).unwrap(),
-            vec![EMPTY_SYMBOL]);
-    }
-
-
-    #[test]
     fn metta_car_atom() {
         let result = run_program("!(eval (car-atom (A $b)))");
         assert_eq!(result, Ok(vec![vec![expr!("A")]]));
@@ -2022,32 +1248,6 @@ mod tests {
         assert_eq!(run_program(&format!("!(cdr-atom ())")), Ok(vec![vec![expr!("Error" ("cdr-atom" ()) {Str::from_str("cdr-atom expects a non-empty expression as an argument")})]]));
         assert_eq!(run_program(&format!("!(cdr-atom a)")), Ok(vec![vec![expr!("Error" ("cdr-atom" "a") {Str::from_str("cdr-atom expects a non-empty expression as an argument")})]]));
         assert_eq!(run_program(&format!("!(cdr-atom $a)")), Ok(vec![vec![expr!("Error" ("cdr-atom" a) {Str::from_str("cdr-atom expects a non-empty expression as an argument")})]]));
-    }
-
-    #[test]
-    fn metta_size_atom() {
-        assert_eq!(run_program(&format!("!(size-atom (5 4 3 2 1))")), Ok(vec![vec![expr!({Number::Integer(5)})]]));
-        assert_eq!(run_program(&format!("!(size-atom ())")), Ok(vec![vec![expr!({Number::Integer(0)})]]));
-    }
-
-    #[test]
-    fn metta_min_atom() {
-        assert_eq!(run_program(&format!("!(min-atom (5 4 5.5))")), Ok(vec![vec![expr!({Number::Integer(4)})]]));
-        assert_eq!(run_program(&format!("!(min-atom ())")), Ok(vec![vec![expr!("Error" ({ MinAtomOp{} } ()) "Empty expression")]]));
-        assert_eq!(run_program(&format!("!(min-atom (3 A B 5))")), Ok(vec![vec![expr!("Error" ({ MinAtomOp{} } ({Number::Integer(3)} "A" "B" {Number::Integer(5)})) "Only numbers are allowed in expression")]]));
-    }
-
-    #[test]
-    fn metta_max_atom() {
-        assert_eq!(run_program(&format!("!(max-atom (5 4 5.5))")), Ok(vec![vec![expr!({Number::Float(5.5)})]]));
-        assert_eq!(run_program(&format!("!(max-atom ())")), Ok(vec![vec![expr!("Error" ({ MaxAtomOp{} } ()) "Empty expression")]]));
-        assert_eq!(run_program(&format!("!(max-atom (3 A B 5))")), Ok(vec![vec![expr!("Error" ({ MaxAtomOp{} } ({Number::Integer(3)} "A" "B" {Number::Integer(5)})) "Only numbers are allowed in expression")]]));
-    }
-
-    #[test]
-    fn metta_index_atom() {
-        assert_eq!(run_program(&format!("!(index-atom (5 4 3 2 1) 2)")), Ok(vec![vec![expr!({Number::Integer(3)})]]));
-        assert_eq!(run_program(&format!("!(index-atom (A B C D E) 5)")), Ok(vec![vec![expr!("Error" ({ IndexAtomOp{} } ("A" "B" "C" "D" "E") {Number::Integer(5)}) "Index is out of bounds")]]));
     }
 
     #[test]
@@ -2719,24 +1919,6 @@ mod tests {
     }
 
     #[test]
-    fn test_error_is_used_as_an_argument() {
-        let metta = Metta::new(Some(EnvBuilder::test_env()));
-        let parser = SExprParser::new(r#"
-            !(get-type Error)
-            !(get-metatype Error)
-            !(get-type (Error Foo Boo))
-            !(Error (+ 1 2) (+ 1 +))
-        "#);
-
-        assert_eq_metta_results!(metta.run(parser), Ok(vec![
-            vec![expr!("->" "Atom" "Atom" "ErrorType")],
-            vec![expr!("Symbol")],
-            vec![expr!("ErrorType")],
-            vec![expr!("Error" ({SumOp{}} {Number::Integer(1)} {Number::Integer(2)}) ({SumOp{}} {Number::Integer(1)} {SumOp{}}))],
-        ]));
-    }
-
-    #[test]
     fn test_string_parsing() {
         let metta = Metta::new(Some(EnvBuilder::test_env()));
         let parser = SExprParser::new(r#"
@@ -2756,85 +1938,6 @@ mod tests {
             vec![expr!({Str::from_str("te\nst")})],
             vec![expr!({Str::from_str("te\nst")} "test")],
         ]));
-    }
-
-    #[test]
-    fn mod_space_op() {
-        let program = r#"
-            !(bind! &new_space (new-space))
-            !(add-atom &new_space (mod-space! stdlib))
-            !(get-atoms &new_space)
-        "#;
-        let runner = Metta::new(Some(runner::environment::EnvBuilder::test_env()));
-        let result = runner.run(SExprParser::new(program)).unwrap();
-
-        let stdlib_space = runner.module_space(runner.get_module_by_name("stdlib").unwrap());
-        assert_eq!(result[2], vec![Atom::gnd(stdlib_space)]);
-    }
-
-    #[test]
-    fn match_op() {
-        let space = DynSpace::new(metta_space("(A B)"));
-        let match_op = MatchOp{};
-        assert_eq!(match_op.execute(&mut vec![expr!({space}), expr!("A" "B"), expr!("B" "A")]),
-            Ok(vec![expr!("B" "A")]));
-    }
-
-    #[test]
-    fn match_op_issue_530() {
-        let space = DynSpace::new(metta_space("(A $a $a)"));
-        let match_op = MatchOp{};
-        let result = match_op.execute(&mut vec![expr!({space}), expr!("A" x y), expr!("A" x y)]).unwrap();
-        assert_eq!(result.len(), 1);
-        assert!(atoms_are_equivalent(&result[0], &expr!("A" x x)),
-            "atoms are not equivalent: expected: {}, actual: {}", expr!("A" x x), result[0]);
-    }
-
-
-    #[test]
-    fn new_space_op() {
-        let res = NewSpaceOp{}.execute(&mut vec![]).expect("No result returned");
-        let space = res.get(0).expect("Result is empty");
-        let space = space.as_gnd::<DynSpace>().expect("Result is not space");
-        let space_atoms: Vec<Atom> = space.borrow().as_space().atom_iter().unwrap().cloned().collect();
-        assert_eq_no_order!(space_atoms, Vec::<Atom>::new());
-    }
-
-    #[test]
-    fn add_atom_op() {
-        let space = DynSpace::new(GroundingSpace::new());
-        let satom = Atom::gnd(space.clone());
-        let res = AddAtomOp{}.execute(&mut vec![satom, expr!(("foo" "bar"))]).expect("No result returned");
-        assert_eq!(res, vec![UNIT_ATOM]);
-        let space_atoms: Vec<Atom> = space.borrow().as_space().atom_iter().unwrap().cloned().collect();
-        assert_eq_no_order!(space_atoms, vec![expr!(("foo" "bar"))]);
-    }
-
-    #[test]
-    fn remove_atom_op() {
-        let space = DynSpace::new(metta_space("
-            (foo bar)
-            (bar foo)
-        "));
-        let satom = Atom::gnd(space.clone());
-        let res = RemoveAtomOp{}.execute(&mut vec![satom, expr!(("foo" "bar"))]).expect("No result returned");
-        // REM: can return Bool in future
-        assert_eq!(res, vec![UNIT_ATOM]);
-        let space_atoms: Vec<Atom> = space.borrow().as_space().atom_iter().unwrap().cloned().collect();
-        assert_eq_no_order!(space_atoms, vec![expr!(("bar" "foo"))]);
-    }
-
-    #[test]
-    fn get_atoms_op() {
-        let space = DynSpace::new(metta_space("
-            (foo bar)
-            (bar foo)
-        "));
-        let satom = Atom::gnd(space.clone());
-        let res = GetAtomsOp{}.execute(&mut vec![satom]).expect("No result returned");
-        let space_atoms: Vec<Atom> = space.borrow().as_space().atom_iter().unwrap().cloned().collect();
-        assert_eq_no_order!(res, space_atoms);
-        assert_eq_no_order!(res, vec![expr!(("foo" "bar")), expr!(("bar" "foo"))]);
     }
 
     #[test]
@@ -2895,104 +1998,6 @@ mod tests {
     }
 
     #[test]
-    fn unique_op() {
-        let unique_op = UniqueAtomOp{};
-        let actual = unique_op.execute(&mut vec![expr!(
-                ("A" ("B" "C"))
-                ("A" ("B" "C"))
-                ("f" "g")
-                ("f" "g")
-                ("f" "g")
-                "Z"
-        )]).unwrap();
-        assert_eq_no_order!(actual,
-                   vec![expr!(("A" ("B" "C")) ("f" "g") "Z")]);
-    }
-
-    #[test]
-    fn union_op() {
-        let union_op = UnionAtomOp{};
-        let actual = union_op.execute(&mut vec![expr!(
-                ("A" ("B" "C"))
-                ("A" ("B" "C"))
-                ("f" "g")
-                ("f" "g")
-                ("f" "g")
-                "Z"
-            ), expr!(
-                ("A" ("B" "C"))
-                "p"
-                "p"
-                ("Q" "a")
-            )]).unwrap();
-        assert_eq_no_order!(actual,
-                   vec![expr!(("A" ("B" "C")) ("A" ("B" "C"))
-                        ("f" "g") ("f" "g") ("f" "g") "Z"
-                        ("A" ("B" "C")) "p" "p" ("Q" "a"))]);
-    }
-
-    #[test]
-    fn intersection_op() {
-        let intersection_op = IntersectionAtomOp{};
-        let actual = intersection_op.execute(&mut vec![expr!(
-                "Z"
-                ("A" ("B" "C"))
-                ("A" ("B" "C"))
-                ("f" "g")
-                ("f" "g")
-                ("f" "g")
-                ("P" "b")
-            ), expr!(
-                ("f" "g")
-                ("f" "g")
-                ("A" ("B" "C"))
-                "p"
-                "p"
-                ("Q" "a")
-                "Z"
-            )]).unwrap();
-        assert_eq_no_order!(actual, vec![expr!("Z" ("A" ("B" "C")) ("f" "g") ("f" "g"))]);
-
-        assert_eq_no_order!(intersection_op.execute(&mut vec![expr!(
-                { Number::Integer(5) }
-                { Number::Integer(4) }
-                { Number::Integer(3) }
-                { Number::Integer(2) }
-            ), expr!(
-                { Number::Integer(5) }
-                { Number::Integer(3) }
-            )]).unwrap(), vec![expr!({Number::Integer(5)} {Number::Integer(3)})]);
-    }
-
-    #[test]
-    fn subtraction_op() {
-        let subtraction_op = SubtractionAtomOp{};
-        let actual = subtraction_op.execute(&mut vec![expr!(
-                "Z"
-                "S"
-                "S"
-                ("A" ("B" "C"))
-                ("A" ("B" "C"))
-                ("f" "g")
-                ("f" "g")
-                ("f" "g")
-                ("P" "b")
-            ), expr!(
-                ("f" "g")
-                ("A" ("B" "C"))
-                "p"
-                "P"
-                ("Q" "a")
-                "Z"
-                "S"
-                "S"
-                "S"
-            )]).unwrap();
-        assert_eq_no_order!(actual,
-                   vec![expr!(("A" ("B" "C")) ("f" "g") ("f" "g") ("P" "b"))]);
-    }
-
-    #[test]
     fn println_op() {
         assert_eq!(PrintlnOp{}.execute(&mut vec![sym!("A")]), unit_result());
     }
@@ -3003,10 +2008,7 @@ mod tests {
                    Ok(vec![sym!("42")]));
     }
 
-    #[test]
-    fn nop_op() {
-        assert_eq!(NopOp{}.execute(&mut vec![]), unit_result());
-    }
+
 
     #[test]
     fn let_op_keep_variables_equalities_issue290() {
@@ -3041,19 +2043,6 @@ mod tests {
             ;; [(→ P R)]
         ";
         assert_eq_metta_results!(run_program(program), Ok(vec![vec![expr!("→" "P" "R")]]));
-    }
-
-    #[test]
-    fn state_ops() {
-        let result = NewStateOp{}.execute(&mut vec![expr!("A" "B")]).unwrap();
-        let old_state = result.get(0).ok_or("error").unwrap();
-        assert_eq!(old_state, &Atom::gnd(StateAtom::new(expr!("A" "B"))));
-        let result = ChangeStateOp{}.execute(&mut vec!(old_state.clone(), expr!("C" "D"))).unwrap();
-        let new_state = result.get(0).ok_or("error").unwrap();
-        assert_eq!(old_state, new_state);
-        assert_eq!(new_state, &Atom::gnd(StateAtom::new(expr!("C" "D"))));
-        let result = GetStateOp{}.execute(&mut vec![new_state.clone()]);
-        assert_eq!(result, Ok(vec![expr!("C" "D")]))
     }
 
     #[test]
@@ -3134,39 +2123,5 @@ mod tests {
         assert_eq!(res, Err(ExecError::from("Range is empty")));
     }
 
-    #[test]
-    fn size_atom_op() {
-        let res = SizeAtomOp{}.execute(&mut vec![expr!({Number::Integer(5)} {Number::Integer(4)} {Number::Integer(3)} {Number::Integer(2)} {Number::Integer(1)})]).expect("No result returned");
-        assert_eq!(res, vec![expr!({Number::Integer(5)})]);
-        let res = SizeAtomOp{}.execute(&mut vec![expr!()]).expect("No result returned");
-        assert_eq!(res, vec![expr!({Number::Integer(0)})]);
-    }
 
-    #[test]
-    fn min_atom_op() {
-        let res = MinAtomOp{}.execute(&mut vec![expr!({Number::Integer(5)} {Number::Integer(4)} {Number::Float(5.5)})]).expect("No result returned");
-        assert_eq!(res, vec![expr!({Number::Integer(4)})]);
-        let res = MinAtomOp{}.execute(&mut vec![expr!({Number::Integer(5)} {Number::Integer(4)} "A")]);
-        assert_eq!(res, Err(ExecError::from("Only numbers are allowed in expression")));
-        let res = MinAtomOp{}.execute(&mut vec![expr!()]);
-        assert_eq!(res, Err(ExecError::from("Empty expression")));
-    }
-
-    #[test]
-    fn max_atom_op() {
-        let res = MaxAtomOp{}.execute(&mut vec![expr!({Number::Integer(5)} {Number::Integer(4)} {Number::Float(5.5)})]).expect("No result returned");
-        assert_eq!(res, vec![expr!({Number::Float(5.5)})]);
-        let res = MaxAtomOp{}.execute(&mut vec![expr!({Number::Integer(5)} {Number::Integer(4)} "A")]);
-        assert_eq!(res, Err(ExecError::from("Only numbers are allowed in expression")));
-        let res = MaxAtomOp{}.execute(&mut vec![expr!()]);
-        assert_eq!(res, Err(ExecError::from("Empty expression")));
-    }
-
-    #[test]
-    fn index_atom_op() {
-        let res = IndexAtomOp{}.execute(&mut vec![expr!({Number::Integer(5)} {Number::Integer(4)} {Number::Integer(3)} {Number::Integer(2)} {Number::Integer(1)}), expr!({Number::Integer(2)})]).expect("No result returned");
-        assert_eq!(res, vec![expr!({Number::Integer(3)})]);
-        let res = IndexAtomOp{}.execute(&mut vec![expr!({Number::Integer(5)} {Number::Integer(4)} {Number::Integer(3)} {Number::Integer(2)} {Number::Integer(1)}), expr!({Number::Integer(5)})]);
-        assert_eq!(res, Err(ExecError::from("Index is out of bounds")));
-    }
 }
